fix(core, android): Fix user score and rating type

+++ application/android/core/src/main/java/com/makentoshe/habrachan/application/android/database/record/UserRecord.kt
@@ -14,9 +14,9 @@ data class UserRecord(
     val speciality: String,
 //    val gender: Int TODO make enum later
     val timeRegisteredRaw: String,
-    val rating: Int,
+    val rating: Float,
     val ratingPosition: Int,
-    val score: Int,
+    val score: Float,
     val postsCount: Int,
     val commentsCount: Int,
     val favoritesCount: Int,
